Safari-609.1.11.1/LayoutTests/fast/selectors/querySelector-with-id-in-quirks-mode-expected.txt

Test #ID selector matching in quirks mode

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS root.querySelectorAll("#Cocoa").length is 1
PASS root.querySelectorAll("#Cocoa")[0].getAttribute("data") is "pass1"
PASS root.querySelectorAll("#Cappuccino").length is 1
PASS root.querySelectorAll("#Cappuccino")[0].getAttribute("data") is "pass2"
PASS root.querySelectorAll("#Cocoa > div").length is 1
PASS root.querySelectorAll("#Cocoa > div")[0].getAttribute("data") is "pass2"
PASS root.querySelectorAll("#Matcha").length is 2
PASS root.querySelectorAll("#Matcha")[0].getAttribute("data") is "pass3"
PASS root.querySelectorAll("#Matcha")[1].getAttribute("data") is "pass4"
PASS root.querySelectorAll("#Matcha > div").length is 1
PASS root.querySelectorAll("#Matcha > div")[0].getAttribute("data") is "pass4"
PASS successfullyParsed is true

TEST COMPLETE
